r/frontdoor_firewall_policy: moving the validation tests over
